Refactor: move sensory primitives from voice.py to perception.py
can_see/can_hear/can_smell + their capacity/condition helpers,
thresholds and override-condition constants were in voice.py only for
historical reasons (born for the speech-attribution chain). They are
general perception primitives -> moved to perception.py beside
can_perceive_sense/blocked_senses.

- perception.py now owns the primitives and imports SIGHT_OVERRIDE_CONDITION
  from combat.capacity (a leaf, no cycle).
- voice.py imports can_see/can_hear/_read_capacity back from perception;
  perception no longer imports voice (deps point one way now:
  voice -> perception -> combat.capacity).
- Repointed: emote.py, tests (blindsight, capacity_chrome, voice_identity).

Pure relocation, no behavior change. 195 tests green.

Closes #637
